% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_simulate_data.R
\name{ef_simulateData}
\alias{ef_simulateData}
\title{Simulate Election data}
\usage{
ef_simulateData(n = 2000, nCov = 0, model)
}
\arguments{
\item{n}{integer, the number of sample points representing observations of election units (e.g., precinct, ballot boxes)}

\item{nCov}{number of covariates affecting both turnout and votes for the winner in each election unit}

\item{model}{see \code{\link{eforensics}}}
}
\value{
The function returns a list with a data frame with the simulated data and a sublist with the parameters used to generate the data
}
\description{
This function simulates data sets from the Election Forensics Finite Mixture models.
}
