% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_main.R
\name{eforensics}
\alias{eforensics}
\title{Election Forensics Finite Mixture Model}
\usage{
eforensics(formula1, formula2, data, weights, mcmc, model, parameters = NULL,
  na.action = "exclude")
}
\arguments{
\item{formula1}{an object of the class \code{formula} as used in \code{\link{lm}}. The independent variable must the votes for the winner party or candidate, whose format must be either integers (number of votes) if the model used is the binomial model, or proportions (of votes) if the model used is the restricted normal (see \code{model})}

\item{formula2}{an object of the class \code{formula} as used in \code{\link{lm}}. The independent variable must the abstention , whose format must be same as the independent variable in \code{formula1}}

\item{data}{a dara.frame with the independent variables (voters for the winner and abstention) and the covariates. If the independent variables are counts, it must contain also a variable \code{N} which must be an integer with the number of elegible voters}

\item{weights}{(not used)}

\item{mcmc}{a list containing \code{n.iter}, which is the number of iterations for the MCMC, \code{burn.in} for the burn-in period of the MCMC chain, \code{n.adapt} indicating the number of adaptative steps before the estimation (see \code{\link{rjags}})}

\item{model}{a string with either "bl", "rn", "rn_no_alpha". It indicates the forensics model of the data. "rn" stands for restricted normal model. "bl" stands binomial with logistic transformation of the probability parameter. These options indicate which model should be used as the distribution for the voters for the winner and abstention.}

\item{parameters}{a string vector with the names of the parameters to monitor. When \code{NULL} (default), it will monitor all the parameters, except the Z's. When \code{parameters='all'}, it will monitor all parameters, including Z, which is necessary to classify the observations as fraudulent cases or not.}

\item{na.action}{(not used)}
}
\value{
The function returns a nested list. The first element of the list is a \code{mcmc} object with the samples from the posterior distribution. The second element of the list is a list of summaries (HPD, Mean, etc)
}
\description{
This function estimates a finite mixture model of election fraud
}
